Модернизировал добавление в список продуктов в закупке. Реализовал удаление товаров, если закупка отменяется

+++ CurrentProduct.cs
@@ -16,7 +16,8 @@ namespace Trading_Company
     {
         int id, id_purchase, count, price;
         bool flag = false;
-        byte[] photo;
+        byte[] photo, basePhoto;
+        string name, category, vat, unit, summary;
         public CurrentProduct(int id, int id_purchase, int count, int price)
         {
             InitializeComponent();
@@ -29,15 +30,21 @@ namespace Trading_Company
 
             SQLiteCommand param = new SQLiteCommand("SELECT name FROM products WHERE id = '"+ id +"'", ConnectionToDB.DB);
             nameBox.Text = param.ExecuteScalar().ToString();
+            name = nameBox.Text;
 
             param = new SQLiteCommand("SELECT category FROM products WHERE id = '" + id + "'", ConnectionToDB.DB);
             categoryBox.Text = param.ExecuteScalar().ToString();
+            category = categoryBox.Text;
 
             param = new SQLiteCommand("SELECT vat FROM products WHERE id = '" + id + "'", ConnectionToDB.DB);
             vatBox.Text = param.ExecuteScalar().ToString();
+            vat = vatBox.Text;
 
             param = new SQLiteCommand("SELECT unit_of_measurement FROM products WHERE id = '" + id + "'", ConnectionToDB.DB);
             unitBox.Text = param.ExecuteScalar().ToString();
+            unit = unitBox.Text;
+
+            summary = richSummaryBox.Text;
 
             countBox.Text = count.ToString();
 
@@ -48,6 +55,7 @@ namespace Trading_Company
             byte[] image = (byte[])param.ExecuteScalar();
             pictureBox1.Image = byteArrayToImage(image);
             photo = image;
+            basePhoto = photo;
 
             param = new SQLiteCommand("SELECT characteristic FROM products WHERE id = '" + id + "'", ConnectionToDB.DB);
             richSummaryBox.Text = param.ExecuteScalar().ToString();
@@ -80,11 +88,11 @@ namespace Trading_Company
                 {
                     //Get the path of specified file
                     filePath = openFileDialog.FileName;
+                    pictureBox1.Image = Image.FromFile(filePath);
 
+                    photo = GetPhoto(filePath);
                 }
-                pictureBox1.Image = Image.FromFile(filePath);
 
-                photo = GetPhoto(filePath);
             }
         }
 
@@ -104,58 +112,62 @@ namespace Trading_Company
         {
             if (nameBox.Text != "" && categoryBox.Text != "" && unitBox.Text != "" && vatBox.Text != "" && countBox.Text != "" && priceBox.Text != "")
             {
-                ConnectionToDB.openDB();
-
-
-                // А если пользователь ничего не изменил?!?!?!?!
-
-                SQLiteCommand ifCMD = new SQLiteCommand("SELECT count_product FROM warehouse WHERE id_product = '" + id + "'", ConnectionToDB.DB);
-
-                if (Convert.ToInt32(ifCMD.ExecuteScalar()) == count)
+                if (!(nameBox.Text.Equals(name) && categoryBox.Text.Equals(category) && unitBox.Text.Equals(unit) && vatBox.Text.Equals(vat) && countBox.Text.Equals(count.ToString()) && priceBox.Text.Equals(price.ToString()) && richSummaryBox.Text.Equals(summary) && Enumerable.SequenceEqual(photo, basePhoto)))
                 {
-                    SQLiteCommand delete = new SQLiteCommand("DELETE FROM warehouse WHERE id_product = '" + id + "'", ConnectionToDB.DB);
-                    delete.ExecuteNonQuery();
-                    delete = new SQLiteCommand("DELETE FROM products WHERE id = '" + id + "'", ConnectionToDB.DB);
-                    delete.ExecuteNonQuery();
+                    ConnectionToDB.openDB();
 
-                    SQLiteCommand create = new SQLiteCommand("INSERT INTO products (name, category, characteristic, unit_of_measurement, vat, image) VALUES ('" + nameBox.Text + "', '" + categoryBox.Text + "', " +
-                    "'" + richSummaryBox.Text + "', '" + unitBox.Text + "', '" + vatBox.Text + "', @photo)", ConnectionToDB.DB);
-                    create.Parameters.AddWithValue("@photo", photo);
-                    create.ExecuteNonQuery();
 
-                    SQLiteCommand getID = new SQLiteCommand("SELECT MAX(id) FROM products", ConnectionToDB.DB);
-                    int id_prod = Convert.ToInt32(getID.ExecuteScalar());
+                    // А если пользователь ничего не изменил?!?!?!?!
 
-                    create = new SQLiteCommand("INSERT INTO warehouse(id_product, count_product, sector_on_warehouse) VALUES ('" + id_prod + "', '" + Convert.ToInt32(countBox.Text) + "', '" + getID.ExecuteScalar().ToString() + nameBox.Text[0] + "')", ConnectionToDB.DB);
-                    create.ExecuteNonQuery();
+                    SQLiteCommand ifCMD = new SQLiteCommand("SELECT count_product FROM warehouse WHERE id_product = '" + id + "'", ConnectionToDB.DB);
 
-                    SQLiteCommand update = new SQLiteCommand("UPDATE product_list_in_purchase SET id_product = '" + id_prod + "', price_product_in_purchase = '"+ Convert.ToInt32(priceBox.Text) + "', count_product_in_purchase = '" + Convert.ToInt32(countBox.Text) + "' WHERE id_product = '"+ id + "' AND id_purchase = '" + id_purchase + "' AND count_product_in_purchase = '" + count + "' AND price_product_in_purchase = '" + price + "'", ConnectionToDB.DB);
-                    update.ExecuteNonQuery();
+                    if (Convert.ToInt32(ifCMD.ExecuteScalar()) == count)
+                    {
+                        SQLiteCommand delete = new SQLiteCommand("DELETE FROM warehouse WHERE id_product = '" + id + "'", ConnectionToDB.DB);
+                        delete.ExecuteNonQuery();
+                        delete = new SQLiteCommand("DELETE FROM products WHERE id = '" + id + "'", ConnectionToDB.DB);
+                        delete.ExecuteNonQuery();
+
+                        SQLiteCommand create = new SQLiteCommand("INSERT INTO products (name, category, characteristic, unit_of_measurement, vat, image) VALUES ('" + nameBox.Text + "', '" + categoryBox.Text + "', " +
+                        "'" + richSummaryBox.Text + "', '" + unitBox.Text + "', '" + vatBox.Text + "', @photo)", ConnectionToDB.DB);
+                        create.Parameters.AddWithValue("@photo", photo);
+                        create.ExecuteNonQuery();
+
+                        SQLiteCommand getID = new SQLiteCommand("SELECT MAX(id) FROM products", ConnectionToDB.DB);
+                        int id_prod = Convert.ToInt32(getID.ExecuteScalar());
+
+                        create = new SQLiteCommand("INSERT INTO warehouse(id_product, count_product, sector_on_warehouse) VALUES ('" + id_prod + "', '" + Convert.ToInt32(countBox.Text) + "', '" + getID.ExecuteScalar().ToString() + nameBox.Text[0] + "')", ConnectionToDB.DB);
+                        create.ExecuteNonQuery();
+
+                        SQLiteCommand update = new SQLiteCommand("UPDATE product_list_in_purchase SET id_product = '" + id_prod + "', price_product_in_purchase = '" + Convert.ToInt32(priceBox.Text) + "', count_product_in_purchase = '" + Convert.ToInt32(countBox.Text) + "' WHERE id_product = '" + id + "' AND id_purchase = '" + id_purchase + "' AND count_product_in_purchase = '" + count + "' AND price_product_in_purchase = '" + price + "'", ConnectionToDB.DB);
+                        update.ExecuteNonQuery();
+                    }
+                    else
+                    {
+                        SQLiteCommand delete = new SQLiteCommand("UPDATE FROM warehouse SET count_product = count_product - '" + count + "'", ConnectionToDB.DB);
+                        delete.ExecuteNonQuery();
+
+                        SQLiteCommand create = new SQLiteCommand("INSERT INTO products (name, category, characteristic, unit_of_measurement, vat, image) VALUES ('" + nameBox.Text + "', '" + categoryBox.Text + "', " +
+                            "'" + richSummaryBox.Text + "', '" + unitBox.Text + "', '" + vatBox.Text + "', @photo", ConnectionToDB.DB);
+                        create.Parameters.AddWithValue("@photo", photo);
+                        create.ExecuteNonQuery();
+
+                        SQLiteCommand getID = new SQLiteCommand("SELECT id(MAX) FROM products", ConnectionToDB.DB);
+                        int id_prod = Convert.ToInt32(getID.ExecuteScalar());
+
+                        create = new SQLiteCommand("INSERT INTO warehouse(id_product, count_product, sector_on_warehouse) VALUES ('" + id_prod + "', '" + Convert.ToInt32(countBox.Text) + "', '" + getID.ExecuteScalar().ToString() + nameBox.Text[0] + "')", ConnectionToDB.DB);
+                        create.ExecuteNonQuery();
+
+                        SQLiteCommand update = new SQLiteCommand("UPDATE FROM product_list_in_purchase SET id_product = '" + id_prod + "', price_product_in_purchase = '" + Convert.ToInt32(priceBox.Text) + "', count_product_in_purchase = '" + Convert.ToInt32(countBox.Text) + "' WHERE id_product = '" + id + "' AND id_purchase = '" + id_purchase + "' AND count_product_in_purchase = '" + count + "' AND price_product_in_purchase = '" + price + "'", ConnectionToDB.DB);
+                        update.ExecuteNonQuery();
+                    }
+
+                    ConnectionToDB.closeDB();
                 }
                 else
                 {
-                    SQLiteCommand delete = new SQLiteCommand("UPDATE FROM warehouse SET count_product = count_product - '" + count + "'", ConnectionToDB.DB);
-                    delete.ExecuteNonQuery();
-
-                    SQLiteCommand create = new SQLiteCommand("INSERT INTO products (name, category, characteristic, unit_of_measurement, vat, image) VALUES ('" + nameBox.Text + "', '" + categoryBox.Text + "', " +
-                        "'" + richSummaryBox.Text + "', '" + unitBox.Text + "', '" + vatBox.Text + "', @photo", ConnectionToDB.DB);
-                    create.Parameters.AddWithValue("@photo", photo);
-                    create.ExecuteNonQuery();
-
-                    SQLiteCommand getID = new SQLiteCommand("SELECT id(MAX) FROM products", ConnectionToDB.DB);
-                    int id_prod = Convert.ToInt32(getID.ExecuteScalar());
-
-                    create = new SQLiteCommand("INSERT INTO warehouse(id_product, count_product, sector_on_warehouse) VALUES ('" + id_prod + "', '" + Convert.ToInt32(countBox.Text) + "', '" + getID.ExecuteScalar().ToString() + nameBox.Text[0] + "')", ConnectionToDB.DB);
-                    create.ExecuteNonQuery();
-
-                    SQLiteCommand update = new SQLiteCommand("UPDATE FROM product_list_in_purchase SET id_product = '" + id_prod + "', price_product_in_purchase = '" + Convert.ToInt32(priceBox.Text) + "', count_product_in_purchase = '" + Convert.ToInt32(countBox.Text) + "' WHERE id_product = '" + id + "' AND id_purchase = '" + id_purchase + "' AND count_product_in_purchase = '" + count + "' AND price_product_in_purchase = '" + price + "'", ConnectionToDB.DB);
-                    update.ExecuteNonQuery();
+                    MessageBox.Show("Изменений не нашлось!", "Ошибка", MessageBoxButtons.OK, MessageBoxIcon.Error);
                 }
-
-                ConnectionToDB.closeDB();
-                
-
-                
             }
             else
             {
